requirements.txt
supabase
python-dotenv



